Break out query result types into different categories
